+++ src/main/java/edu/hsbremen/cloud/comparison/IComparisonStrategy.java
@@ -1,5 +1,5 @@
 package edu.hsbremen.cloud.comparison;
 
 public interface IComparisonStrategy {
-    Integer compare(byte[] imageReference, byte[] imageCompared);
+    Double compare(byte[] imageReference, byte[] imageCompared);
 }
