+++ src/main/java/edu/hsbremen/cloud/dto/ComparsionDto.java
@@ -3,9 +3,9 @@ package edu.hsbremen.cloud.dto;
 public class ComparsionDto {
     private ImageDto referenceImage;
     private ImageDto comparedImage;
-    private Integer score;
+    private Double score;
 
-    public ComparsionDto(ImageDto referenceImage, ImageDto comparedImage, Integer score) {
+    public ComparsionDto(ImageDto referenceImage, ImageDto comparedImage, Double score) {
         this.referenceImage = referenceImage;
         this.comparedImage = comparedImage;
         this.score = score;
@@ -19,7 +19,7 @@ public class ComparsionDto {
         return comparedImage;
     }
 
-    public Integer getScore() {
+    public Double getScore() {
         return score;
     }
 }
